+++ service/task_billing_test.go
@@ -77,7 +77,7 @@ func truncate(t *testing.T) {
 	})
 }
 
-func seedUser(t *testing.T, id int, quota int) {
+func seedUser(t *testing.T, id int, quota int64) {
 	t.Helper()
 	user := &model.User{Id: id, Username: "test_user", Quota: quota, Status: common.UserStatusEnabled}
 	require.NoError(t, model.DB.Create(user).Error)
@@ -262,7 +262,7 @@ func getUserQuota(t *testing.T, id int) int {
 	t.Helper()
 	var user model.User
 	require.NoError(t, model.DB.Select("quota").Where("id = ?", id).First(&user).Error)
-	return user.Quota
+	return int(user.Quota)
 }
 
 func getUserUsageAccounting(t *testing.T, id int) (int, int) {
